Remove ship velocity

+++ source/GameGraphRoot.cpp
@@ -16,10 +16,10 @@ using namespace std;
 constexpr unsigned int SCENE_WIDTH = 1024;
 
 /** 2 pi */
-constexpr float TWO_PI = 2 * M_PI;
+constexpr float TWO_PI = (float)(2 * M_PI);
 
 /** Pi over 180 for converting between degrees and radians */
-constexpr float PI_180 = M_PI / 180;
+constexpr float PI_180 = (float)(M_PI / 180);
 
 /** The scale of the breach textures. */
 constexpr float BREACH_SCALE = 0.25;
